// src/api/generated.ts
export interface TodoResponse {
  id: string;
  title: string;
  userId: string;
  completed: boolean;
  createdAt: string;
  description?: string;
}